Effective implementation of to_base and from_base
Reimplemented the from_base and to_base using boost fusion. This increases code reusability and removed the preprocessor macro. This way the code is better structured and debuggable.

+++ src/Bun.cpp
@@ -8,7 +8,7 @@ namespace test {
   struct Person {
     std::string name;
     int age;
-    float height;
+    double height;
   };
 }
